Added User.Read into scopes for Connect-Graph
The cmdlet Get-MgOrganization cannot be called without the proper permissions. User.Read is the lowest level permission to run the cmdlet. It's not required to specify User.Read when connecting to MS Graph, but the permission needs to be consented to. By directly specifying this permission in the scopes when connecting, the user will be prompted to grant consent for the permission.